refactor: Move non-selector functions into new helper file
